SQL Delight is working

+++ shared/src/commonMain/kotlin/com/matrix/shared/data/contracts/SmiteRepository.kt
@@ -2,18 +2,18 @@ package com.matrix.shared.data.contracts
 
 import com.matrix.shared.data.model.builds.BuildInformation
 import com.matrix.shared.data.model.gods.GodInformation
-import com.matrix.shared.data.model.skins.GodSkinInformation
 import com.matrix.shared.data.model.items.ItemInformation
+import com.matrix.shared.data.model.skins.GodSkinInformation
 import kotlinx.coroutines.flow.Flow
 
 interface SmiteRepository {
   fun getGods(): Flow<List<GodInformation>>
-  fun getGod(godId: Int): Flow<GodInformation>
-  fun getGodSkins(godId: Int): Flow<List<GodSkinInformation>>
+  fun getGod(godId: Long): Flow<GodInformation>
+  fun getGodSkins(godId: Long): Flow<List<GodSkinInformation>>
   fun getItems(): Flow<List<ItemInformation>>
-  fun getItem(itemId: Int): Flow<ItemInformation>
+  fun getItem(itemId: Long): Flow<ItemInformation>
   fun getBuilds(): Flow<List<BuildInformation>>
-  fun getBuild(buildId: Int): Flow<BuildInformation>
+  fun getBuild(buildId: Long): Flow<BuildInformation>
   suspend fun createBuild(buildInformation: BuildInformation)
   suspend fun deleteBuild(buildInformation: BuildInformation)
   suspend fun sync()
